+++ vec120/src/main/java/com/foursoft/harness/vec/v12x/navigations/CustomPropertyNavs.java
@@ -23,10 +23,10 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.vecmodel.vec120.navigations;
+package com.foursoft.harness.vec.v12x.navigations;
 
+import com.foursoft.harness.vec.v12x.*;
 import com.foursoft.vecmodel.common.HasCustomProperties;
-import com.foursoft.vecmodel.vec120.*;
 
 import java.math.BigInteger;
 import java.util.Collections;
