+++ vec120/src/main/java/com/foursoft/harness/vec/v12x/visitor/ReferencedNodeLocationVisitor.java
@@ -10,10 +10,10 @@
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,9 +23,9 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.vecmodel.vec120.visitor;
+package com.foursoft.harness.vec.v12x.visitor;
 
-import com.foursoft.vecmodel.vec120.*;
+import com.foursoft.harness.vec.v12x.*;
 
 public class ReferencedNodeLocationVisitor extends StrictBaseVisitor<VecTopologyNode> {
